refactor: renamed LoadingImageFailedIndicator file to ImageLoadingFailedIndicator component
